make stand alon screen for sign-in and sign-up
