+++ src/App.js
@@ -3,8 +3,8 @@ import React, { useEffect, useState } from "react";
 import { Route, Routes, Navigate } from "react-router-dom";
 import Navbar from "./components/Navbar";
 import Home from "./pages/Home";
-import SignIn from "./pages/SignIn";
-import SignUp from "./pages/SignUp";
+import SignIn from "../src/screen/signIn";
+import SignUp from "../src/screen/signUp";
 import Books from "./pages/Books";
 import Settings from "./pages/Settings";
 import "./App.css";
