+++ src/screen/signIn/style.js
@@ -1,0 +1,27 @@
+// SignIn/styles.js
+export const signInStyles = {
+  boxContainer: {
+    boxShadow: 3,
+    borderRadius: 2,
+    px: 4,
+    py: 2,
+    marginTop: 8,
+    display: "flex",
+    flexDirection: "column",
+    alignItems: "center",
+  },
+  title: {
+    fontSize: "2rem",
+    fontWeight: "bold",
+    color: (theme) => theme.palette.grey.main,
+    textShadow: "2px 2px 4px rgba(0, 0, 0, 0.2)",
+    marginBottom: "20px",
+  },
+  form: {
+    mt: 1,
+  },
+  button: {
+    mt: 3,
+    mb: 2,
+  },
+};
